TANGERINE-568
git-svn-id: svn://svn.orangeleap.com/tangerine/tangerine/trunk@6124 0f581344-5646-3a48-8ded-616bfea576c6

+++ src/com/orangeleap/tangerine/domain/customization/FieldDefinition.java
@@ -15,7 +15,7 @@ public class FieldDefinition implements Serializable {
     private static final long serialVersionUID = 1L;
 
     private String id;
-    private int numericId;
+    private long numericId;
     private EntityType entityType;
     private ReferenceType referenceType;
     private String fieldName;
@@ -122,11 +122,11 @@ public class FieldDefinition implements Serializable {
                 append("entityType", entityType).append("referenceType", referenceType).append("entityAttributes", entityAttributes).append("site", site).toString();
     }
 
-	public void setNumericId(int numericId) {
+	public void setNumericId(long numericId) {
 		this.numericId = numericId;
 	}
 
-	public int getNumericId() {
+	public long getNumericId() {
 		return numericId;
 	}
